refactor: move the new `bq-slider2` implementation to `bq-slider`
