+++ packages/beeq/src/components.d.ts
@@ -22,8 +22,7 @@ import { TNotificationBorderRadius, TNotificationType } from "./components/notif
 import { TProgressBorderShape, TProgressThickness, TProgressType } from "./components/progress/bq-progress.types";
 import { TRadioGroupOrientation } from "./components/radio-group/bq-radio-group.types";
 import { TSideMenuAppearance, TSideMenuSize } from "./components/side-menu/bq-side-menu.types";
-import { TSliderType } from "./components/slider/bq-slider.types";
-import { TSliderType as TSliderType1, TSliderValue } from "./components/slider2/bq-slider.types";
+import { TSliderType, TSliderValue } from "./components/slider/bq-slider.types";
 import { TSpinnerSize, TSpinnerTextPosition } from "./components/spinner/bq-spinner.types";
 import { TStatusType } from "./components/status/bq-status.types";
 import { TStepsSize, TStepsType } from "./components/steps/bq-steps.types";
@@ -50,8 +49,7 @@ export { TNotificationBorderRadius, TNotificationType } from "./components/notif
 export { TProgressBorderShape, TProgressThickness, TProgressType } from "./components/progress/bq-progress.types";
 export { TRadioGroupOrientation } from "./components/radio-group/bq-radio-group.types";
 export { TSideMenuAppearance, TSideMenuSize } from "./components/side-menu/bq-side-menu.types";
-export { TSliderType } from "./components/slider/bq-slider.types";
-export { TSliderType as TSliderType1, TSliderValue } from "./components/slider2/bq-slider.types";
+export { TSliderType, TSliderValue } from "./components/slider/bq-slider.types";
 export { TSpinnerSize, TSpinnerTextPosition } from "./components/spinner/bq-spinner.types";
 export { TStatusType } from "./components/status/bq-status.types";
 export { TStepsSize, TStepsType } from "./components/steps/bq-steps.types";
@@ -914,44 +912,6 @@ export namespace Components {
     }
     interface BqSlider {
         /**
-          * A number representing the delay value applied to bqChange event handler
-         */
-        "debounceTime": number;
-        /**
-          * If `true` slider is disabled
-         */
-        "disabled"?: boolean;
-        /**
-          * A number representing the minimum value between the min and max range selected.
-         */
-        "gap": number;
-        /**
-          * A number representing the max value of the slider.
-         */
-        "max": number;
-        /**
-          * A number representing the min value of the slider.
-         */
-        "min": number;
-        /**
-          * A number representing the step of the slider.
-         */
-        "step": number;
-        /**
-          * It defines the type of slider to display
-         */
-        "type": TSliderType;
-        /**
-          * A number representing the value of the slider.
-         */
-        "value": number | Array<number> | string;
-        /**
-          * If `true` it will display the min and max values
-         */
-        "valueIndicator"?: boolean;
-    }
-    interface BqSlider2 {
-        /**
           * The amount of time, in milliseconds, to wait to trigger the `bqChange` event after each value change.
          */
         "debounceTime": number;
@@ -982,7 +942,7 @@ export namespace Components {
         /**
           * It defines the type of slider to display
          */
-        "type": TSliderType1;
+        "type": TSliderType;
         /**
           * The value of the slider. - If the slider type is `single`, the value is a number. - If the slider type is `range`, the value is an array of two numbers (the first number represents the `min` value and the second number represents the `max` value).
          */
@@ -1431,10 +1391,6 @@ export interface BqSideMenuItemCustomEvent<T> extends CustomEvent<T> {
 export interface BqSliderCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLBqSliderElement;
-}
-export interface BqSlider2CustomEvent<T> extends CustomEvent<T> {
-    detail: T;
-    target: HTMLBqSlider2Element;
 }
 export interface BqStepItemCustomEvent<T> extends CustomEvent<T> {
     detail: T;
@@ -1892,7 +1848,7 @@ declare global {
         new (): HTMLBqSideMenuItemElement;
     };
     interface HTMLBqSliderElementEventMap {
-        "bqChange": { value: number | Array<number> | string; el: HTMLBqSliderElement };
+        "bqChange": { value: Exclude<TSliderValue, string>; el: HTMLBqSliderElement };
         "bqBlur": HTMLBqSliderElement;
         "bqFocus": HTMLBqSliderElement;
     }
@@ -1909,25 +1865,6 @@ declare global {
     var HTMLBqSliderElement: {
         prototype: HTMLBqSliderElement;
         new (): HTMLBqSliderElement;
-    };
-    interface HTMLBqSlider2ElementEventMap {
-        "bqChange": { value: Exclude<TSliderValue, string>; el: HTMLBqSlider2Element };
-        "bqBlur": HTMLBqSlider2Element;
-        "bqFocus": HTMLBqSlider2Element;
-    }
-    interface HTMLBqSlider2Element extends Components.BqSlider2, HTMLStencilElement {
-        addEventListener<K extends keyof HTMLBqSlider2ElementEventMap>(type: K, listener: (this: HTMLBqSlider2Element, ev: BqSlider2CustomEvent<HTMLBqSlider2ElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
-        addEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
-        addEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
-        addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
-        removeEventListener<K extends keyof HTMLBqSlider2ElementEventMap>(type: K, listener: (this: HTMLBqSlider2Element, ev: BqSlider2CustomEvent<HTMLBqSlider2ElementEventMap[K]>) => any, options?: boolean | EventListenerOptions): void;
-        removeEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
-        removeEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
-        removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
-    }
-    var HTMLBqSlider2Element: {
-        prototype: HTMLBqSlider2Element;
-        new (): HTMLBqSlider2Element;
     };
     /**
      * Spinners are designed for users to display data loading.
@@ -2123,7 +2060,6 @@ declare global {
         "bq-side-menu": HTMLBqSideMenuElement;
         "bq-side-menu-item": HTMLBqSideMenuItemElement;
         "bq-slider": HTMLBqSliderElement;
-        "bq-slider2": HTMLBqSlider2Element;
         "bq-spinner": HTMLBqSpinnerElement;
         "bq-status": HTMLBqStatusElement;
         "bq-step-item": HTMLBqStepItemElement;
@@ -3161,56 +3097,6 @@ declare namespace LocalJSX {
     }
     interface BqSlider {
         /**
-          * A number representing the delay value applied to bqChange event handler
-         */
-        "debounceTime"?: number;
-        /**
-          * If `true` slider is disabled
-         */
-        "disabled"?: boolean;
-        /**
-          * A number representing the minimum value between the min and max range selected.
-         */
-        "gap"?: number;
-        /**
-          * A number representing the max value of the slider.
-         */
-        "max"?: number;
-        /**
-          * A number representing the min value of the slider.
-         */
-        "min"?: number;
-        /**
-          * Handler to be called when the slider loses focus
-         */
-        "onBqBlur"?: (event: BqSliderCustomEvent<HTMLBqSliderElement>) => void;
-        /**
-          * Handler to be called when change the value on range inputs
-         */
-        "onBqChange"?: (event: BqSliderCustomEvent<{ value: number | Array<number> | string; el: HTMLBqSliderElement }>) => void;
-        /**
-          * Handler to be called when the slider gets focused
-         */
-        "onBqFocus"?: (event: BqSliderCustomEvent<HTMLBqSliderElement>) => void;
-        /**
-          * A number representing the step of the slider.
-         */
-        "step"?: number;
-        /**
-          * It defines the type of slider to display
-         */
-        "type"?: TSliderType;
-        /**
-          * A number representing the value of the slider.
-         */
-        "value"?: number | Array<number> | string;
-        /**
-          * If `true` it will display the min and max values
-         */
-        "valueIndicator"?: boolean;
-    }
-    interface BqSlider2 {
-        /**
           * The amount of time, in milliseconds, to wait to trigger the `bqChange` event after each value change.
          */
         "debounceTime"?: number;
@@ -3237,15 +3123,15 @@ declare namespace LocalJSX {
         /**
           * Handler to be called when the slider loses focus
          */
-        "onBqBlur"?: (event: BqSlider2CustomEvent<HTMLBqSlider2Element>) => void;
+        "onBqBlur"?: (event: BqSliderCustomEvent<HTMLBqSliderElement>) => void;
         /**
           * Handler to be called when change the value on range inputs
          */
-        "onBqChange"?: (event: BqSlider2CustomEvent<{ value: Exclude<TSliderValue, string>; el: HTMLBqSlider2Element }>) => void;
+        "onBqChange"?: (event: BqSliderCustomEvent<{ value: Exclude<TSliderValue, string>; el: HTMLBqSliderElement }>) => void;
         /**
           * Handler to be called when the slider gets focused
          */
-        "onBqFocus"?: (event: BqSlider2CustomEvent<HTMLBqSlider2Element>) => void;
+        "onBqFocus"?: (event: BqSliderCustomEvent<HTMLBqSliderElement>) => void;
         /**
           * A number representing the step of the slider. ⚠️ Please notice that the value (or list of values if the slider type is `range`) will be rounded to the nearest multiple of `step`.
          */
@@ -3253,7 +3139,7 @@ declare namespace LocalJSX {
         /**
           * It defines the type of slider to display
          */
-        "type"?: TSliderType1;
+        "type"?: TSliderType;
         /**
           * The value of the slider. - If the slider type is `single`, the value is a number. - If the slider type is `range`, the value is an array of two numbers (the first number represents the `min` value and the second number represents the `max` value).
          */
@@ -3677,7 +3563,6 @@ declare namespace LocalJSX {
         "bq-side-menu": BqSideMenu;
         "bq-side-menu-item": BqSideMenuItem;
         "bq-slider": BqSlider;
-        "bq-slider2": BqSlider2;
         "bq-spinner": BqSpinner;
         "bq-status": BqStatus;
         "bq-step-item": BqStepItem;
@@ -3730,7 +3615,6 @@ declare module "@stencil/core" {
             "bq-side-menu": LocalJSX.BqSideMenu & JSXBase.HTMLAttributes<HTMLBqSideMenuElement>;
             "bq-side-menu-item": LocalJSX.BqSideMenuItem & JSXBase.HTMLAttributes<HTMLBqSideMenuItemElement>;
             "bq-slider": LocalJSX.BqSlider & JSXBase.HTMLAttributes<HTMLBqSliderElement>;
-            "bq-slider2": LocalJSX.BqSlider2 & JSXBase.HTMLAttributes<HTMLBqSlider2Element>;
             /**
              * Spinners are designed for users to display data loading.
              */
