+++ packages/beeq/src/components/slider/__tests__/bq-slider.e2e.ts
