+++ packages/beeq/src/components/slider/bq-slider.types.ts
@@ -1,2 +1,4 @@
 export const SLIDER_TYPE = ['single', 'range'] as const;
 export type TSliderType = (typeof SLIDER_TYPE)[number];
+
+export type TSliderValue = string | number | [number, number];
