+++ packages/beeq/src/components/slider/scss/bq-slider.variables.scss
@@ -1,13 +1,11 @@
 /* -------------------------------------------------------------------------- */
-/*                          Slider custom properties                          */
+/*                   Slider2 custom properties                  */
 /* -------------------------------------------------------------------------- */
 
 :host {
   /**
-   * @prop --bq-slider--size: Slider size
+   * @prop --bq-slider--size: The height of the slider track/progress area
    * @prop --bq-slider--border-radius: Slider border radius
-   * @prop --bq-slider--z-index-input: Slider input z-index
-   * @prop --bq-slider--z-index-progress: Slider progress z-index
    * @prop --bq-slider--thumb-size: Slider hover thumb size
    * @prop --bq-slider--progress-color: Slider progress background color
    * @prop --bq-slider--trackarea-color: Slider track background color
@@ -15,8 +13,6 @@
 
   --bq-slider--size: 4px;
   --bq-slider--border-radius: theme('borderRadius.xs');
-  --bq-slider--z-index-input: 2;
-  --bq-slider--z-index-progress: 1;
   --bq-slider--thumb-size: 13px;
   --bq-slider--progress-color: theme('colors.ui.brand');
   --bq-slider--trackarea-color: theme('colors.ui.secondary');
