--- packages/beeq/src/components/slider2/bq-slider.types.ts
@@ -1,4 +1,0 @@
-export const SLIDER_TYPE = ['single', 'range'] as const;
-export type TSliderType = (typeof SLIDER_TYPE)[number];
-
-export type TSliderValue = string | number | [number, number];
