--- packages/beeq/src/components/slider2/scss/bq-slider2.variables.scss
@@ -1,19 +1,0 @@
-/* -------------------------------------------------------------------------- */
-/*                   Slider2 custom properties                  */
-/* -------------------------------------------------------------------------- */
-
-:host {
-  /**
-   * @prop --bq-slider--size: The height of the slider track/progress area
-   * @prop --bq-slider--border-radius: Slider border radius
-   * @prop --bq-slider--thumb-size: Slider hover thumb size
-   * @prop --bq-slider--progress-color: Slider progress background color
-   * @prop --bq-slider--trackarea-color: Slider track background color
-   */
-
-  --bq-slider--size: 4px;
-  --bq-slider--border-radius: theme('borderRadius.xs');
-  --bq-slider--thumb-size: 13px;
-  --bq-slider--progress-color: theme('colors.ui.brand');
-  --bq-slider--trackarea-color: theme('colors.ui.secondary');
-}
